Reverted LATENCY-Division in FlexPowerplant: Test-Scenario Int-Roundings removed -> Floats Logging: Float-Data

+++ src-test/hsoines/oekoflex/display/MerritOrderGraphTest.java
@@ -65,7 +65,7 @@ public class MerritOrderGraphTest {
         }
 
         @Override
-        public int getTotalClearedQuantity() {
+        public float getTotalClearedQuantity() {
             return 4;
         }
